backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pandas==2.1.3
requests==2.31.0
python-dotenv==1.0.0
python-multipart==0.0.6
python-dateutil==2.8.2